🚚 Move toBitmap to where it belongs #42248
