+++ android/src/main/java/cz/ackee/extensions/android/ViewExtensions.kt
@@ -175,24 +175,3 @@ var TextView.drawableBottom: Drawable?
 fun ViewGroup.inflate(@LayoutRes layout: Int, attachToParent: Boolean = true) {
     return context.inflate(layout, this, attachToParent)
 }
-
-/**
- * Convert this Drawable to Bitmap representation. Should take care of every Drawable type
- */
-fun Drawable.toBitmap(): Bitmap {
-    if (this is BitmapDrawable) {
-        return bitmap
-    }
-
-    val bitmap = if (intrinsicWidth <= 0 || intrinsicHeight <= 0) {
-        Bitmap.createBitmap(1, 1, Bitmap.Config.ARGB_8888)
-    } else {
-        Bitmap.createBitmap(intrinsicWidth, intrinsicHeight, Bitmap.Config.ARGB_8888)
-    }
-
-    Canvas(bitmap).apply {
-        setBounds(0, 0, width, height)
-        draw(this)
-    }
-    return bitmap
-}
